backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
